+++ src/main/java/com/SearchFunction/repository/MediaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MediaRepository extends JpaRepository<Media, Integer> {
+public interface MediaRepository extends JpaRepository<Media, Long> {
 
     List<Media> findByGenreContainingIgnoreCase(String genre);
